% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{alexanderson_homogenize}
\alias{alexanderson_homogenize}
\title{Alexanderson test for all available files (which have successfully passed the second fill)}
\usage{
alexanderson_homogenize(data, folder)
}
\arguments{
\item{data}{data and coordinates}

\item{folder}{directory to save the output files}
}
\value{
data and coordinates
}
\description{
Alexanderson test for all available files (which have successfully passed the second fill)
}
