% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{alexanderson_homogenize_data}
\alias{alexanderson_homogenize_data}
\title{Homogenization - Alexanderson test
Reference series, compares and corrects
For each dataset (1870, 1900...)
The 5 most correlated series are chosen using the difference series
A weighted average is made with the 5 series (correlation * data1 + ...) / sum(correlations), which will be the reference series
Alexanderson will give a breakpoint and a ratio value to multiply the older part by... iterate while breakpoints are given
Save statistics on inhomogeneities. Basically, the number of data points changed in each series and the time of the inhomogeneity.
CSV with the number of data points changed and CSV with inhomogeneity point - all x 12 months}
\usage{
alexanderson_homogenize_data(file_data, no_use_series = c())
}
\arguments{
\item{file_data}{path to the data file}

\item{no_use_series}{series that will not be homogenized}
}
\value{
None
}
\description{
Homogenization - Alexanderson test
Reference series, compares and corrects
For each dataset (1870, 1900...)
The 5 most correlated series are chosen using the difference series
A weighted average is made with the 5 series (correlation * data1 + ...) / sum(correlations), which will be the reference series
Alexanderson will give a breakpoint and a ratio value to multiply the older part by... iterate while breakpoints are given
Save statistics on inhomogeneities. Basically, the number of data points changed in each series and the time of the inhomogeneity.
CSV with the number of data points changed and CSV with inhomogeneity point - all x 12 months
}
