% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{apply_ecdf}
\alias{apply_ecdf}
\title{Anomalías de los datos mensuales}
\usage{
apply_ecdf(data)
}
\arguments{
\item{data}{datos mensuales}
}
\value{
anomalías de los datos
}
\description{
Anomalías de los datos mensuales
}
