% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{apply_ecdf_month}
\alias{apply_ecdf_month}
\title{Anomalías de los datos}
\usage{
apply_ecdf_month(data)
}
\arguments{
\item{data}{datos mensuales}
}
\value{
anomalías de los datos
}
\description{
Anomalías de los datos
}
