% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_data_year}
\alias{calc_data_year}
\title{Suma los datos de cada año, para devolver un solo dato anual}
\usage{
calc_data_year(data)
}
\arguments{
\item{data}{matriz de datos}
}
\value{
un dato por año
}
\description{
Suma los datos de cada año, para devolver un solo dato anual
}
