% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_data_year_month_station}
\alias{calc_data_year_month_station}
\title{Returns the slope z for years and stations}
\usage{
calc_data_year_month_station(data, calc_function)
}
\arguments{
\item{data}{station data}

\item{calc_function}{function to use}
}
\value{
list of results
}
\description{
Returns the slope z for years and stations
}
