% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_mkTrend_pval}
\alias{calc_mkTrend_pval}
\title{Returns the p-value calculated by mkTrend or pval0 if pval was NA}
\usage{
calc_mkTrend_pval(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
pval
}
\description{
Returns the p-value calculated by mkTrend or pval0 if pval was NA
}
