% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_mkTrend_slp}
\alias{calc_mkTrend_slp}
\title{Linear regression of the data against years.}
\usage{
calc_mkTrend_slp(data)
}
\arguments{
\item{data}{index}
}
\value{
lm
}
\description{
Linear regression of the data against years.
}
