% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_percentage}
\alias{calc_percentage}
\title{Percentage difference}
\usage{
calc_percentage(datos, years = NA)
}
\arguments{
\item{datos}{data}

\item{years}{years}
}
\value{
percentage
}
\description{
Percentage difference
}
