% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calculate_reconstruction_statistics}
\alias{calculate_reconstruction_statistics}
\title{Calculate reconstruction statistics
- hydroGOF -- statistic per station
- D / MAE / PBIAS / RMSE  - per station and month}
\usage{
calculate_reconstruction_statistics(sim, obs)
}
\arguments{
\item{sim}{filled data}

\item{obs}{original data}
}
\value{
deleted data and input data per station
}
\description{
Calculate reconstruction statistics
- hydroGOF -- statistic per station
- D / MAE / PBIAS / RMSE  - per station and month
}
