% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Salida final con todos los estadíticos, serie regional promedio, tendencias, SPI...}
\usage{
calculate_statistics(data, data_ori)
}
\arguments{
\item{data}{data y coor}

\item{data_ori}{data original}
}
\value{
data and coor
}
\description{
Salida final con todos los estadíticos, serie regional promedio, tendencias, SPI...
}
