% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{calculate_statistics_data}
\alias{calculate_statistics_data}
\title{Calculates statistics for the data
Monthly, seasonal, and annual trends, Trend package, sens.slope function
Add 1 to everything to avoid zeros
Significance, modifiedmk package, bbsmk function
National average series
SPI at scales 3, 12, and 24 for each series, important to ensure operations cannot be reversed}
\usage{
calculate_statistics_data(file_data, data_ori)
}
\arguments{
\item{file_data}{data and coordinates}

\item{data_ori}{original data}
}
\value{
None
}
\description{
Calculates statistics for the data
Monthly, seasonal, and annual trends, Trend package, sens.slope function
Add 1 to everything to avoid zeros
Significance, modifiedmk package, bbsmk function
National average series
SPI at scales 3, 12, and 24 for each series, important to ensure operations cannot be reversed
}
