% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterraneancalculationsNews.R
\name{ClimIndNews}
\alias{ClimIndNews}
\title{mediterraneancalculationsNews}
\usage{
ClimIndNews()
}
\description{
Show the NEWS file of the \pkg{mediterraneancalculations} package.
}
\details{
(See description)
}
