% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{coef_var}
\alias{coef_var}
\title{Coeficientes de variación, desviación estándar
https://fhernanb.github.io/Manual-de-R/varia.html}
\usage{
coef_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{datos}

\item{na.rm}{Ignorara NAs}
}
\value{
percentage
}
\description{
Coeficientes de variación, desviación estándar
https://fhernanb.github.io/Manual-de-R/varia.html
}
