% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{delete_zero}
\alias{delete_zero}
\title{Removes data if there are 8 or more consecutive months of zeros, and if one of the involved months has less than 70\% zeros, its data is removed. Also removes data if there are 5 or more consecutive months of zeros, and if all the involved months have less than 70\% zeros, all are removed.}
\usage{
delete_zero(data)
}
\arguments{
\item{data}{data}
}
\value{
data with zero groups removed
}
\description{
Removes data if there are 8 or more consecutive months of zeros, and if one of the involved months has less than 70\% zeros, its data is removed. Also removes data if there are 5 or more consecutive months of zeros, and if all the involved months have less than 70\% zeros, all are removed.
}
