% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{delete_zones}
\alias{delete_zones}
\title{Below 28 degrees north, remove stations}
\usage{
delete_zones(data)
}
\arguments{
\item{data}{data and coordinates}
}
\value{
data and coordinates
}
\description{
Below 28 degrees north, remove stations
}
