% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{dry_spell_trend}
\alias{dry_spell_trend}
\title{esta función calcula la tendencia. Hay que definirle un objeto de años (years) con el año correspondiente a cada caso.}
\usage{
dry_spell_trend(index, threshold)
}
\arguments{
\item{index}{index}

\item{threshold}{threshold}
}
\value{
output
}
\description{
esta función calcula la tendencia. Hay que definirle un objeto de años (years) con el año correspondiente a cada caso.
}
