% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{fill_one_series}
\alias{fill_one_series}
\title{Rellena la serie recibida utilizando las otras en el orden en el que están en other_series}
\usage{
fill_one_series(series, other_series)
}
\arguments{
\item{series}{serie de datos a completar}

\item{other_series}{series de datos con las que copletar en el orden en el que se tienen que utilizar}
}
\value{
serie de datosd rellena
}
\description{
Rellena la serie recibida utilizando las otras en el orden en el que están en other_series
}
