% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{fill_series}
\alias{fill_series}
\title{Monthly series filling
We use stations less than 200km away with correlations above 0.7.
For June, July, and August, we fill with the closest station.
Use the method that correlates best with the original series.}
\usage{
fill_series(control_data, min_correlation, max_dist)
}
\arguments{
\item{control_data}{data from the stations and their coordinates}

\item{min_correlation}{minimum correlation to use the data in filling}

\item{max_dist}{maximum distance between series to use}
}
\value{
data and coordinates with data that did not pass the control removed
}
\description{
Monthly series filling
We use stations less than 200km away with correlations above 0.7.
For June, July, and August, we fill with the closest station.
Use the method that correlates best with the original series.
}
