% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{fill_unfillable_station}
\alias{fill_unfillable_station}
\title{In countries where no series are found, we allow up to three years of data to be filled with the average. For example, if for a specific period 1900-2020 no series appear, but there are a maximum of three years of data (i.e., 36 months), we fill these data with the average of the 15 closest data points in time.
This applies as long as these five years are not between 2015 and 2020 or within the first five years of the series, as this could affect trends. 
If the series are from 1981-2020, the same rule applies, but only two years of lost data are allowed.}
\usage{
fill_unfillable_station(data, fillable_years)
}
\arguments{
\item{data}{station data to be filled}

\item{fillable_years}{years that can be filled with the station's monthly average}
}
\value{
None
}
\description{
In countries where no series are found, we allow up to three years of data to be filled with the average. For example, if for a specific period 1900-2020 no series appear, but there are a maximum of three years of data (i.e., 36 months), we fill these data with the average of the 15 closest data points in time.
This applies as long as these five years are not between 2015 and 2020 or within the first five years of the series, as this could affect trends. 
If the series are from 1981-2020, the same rule applies, but only two years of lost data are allowed.
}
