% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{main_mediterranean_calculations}
\alias{main_mediterranean_calculations}
\title{Reads precipitation files, calculates statistics, and saves the results
The input files are 2 CSVs: one with coordinates in degrees (stations in rows and lat/lon in columns) and the other with monthly data (dates in rows and stations in columns)}
\usage{
main_mediterranean_calculations(file_data, file_coor)
}
\arguments{
\item{file_data}{path to the data file}

\item{file_coor}{path to the coordinates file}
}
\value{
None
}
\description{
Reads precipitation files, calculates statistics, and saves the results
The input files are 2 CSVs: one with coordinates in degrees (stations in rows and lat/lon in columns) and the other with monthly data (dates in rows and stations in columns)
}
