% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{main_mediterranean_calculations_}
\alias{main_mediterranean_calculations_}
\title{Calculates statistics for a country}
\usage{
main_mediterranean_calculations_(read_all_data, folder, pb = NULL)
}
\arguments{
\item{read_all_data}{input data}

\item{folder}{directory where files are saved}

\item{pb}{progress bar}
}
\value{
None
}
\description{
Calculates statistics for a country
}
