% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{mediterranean_calculations}
\alias{mediterranean_calculations}
\title{Hace el control de calidad
Control de calidad: Se estaciones con menos de 20 años de datos y usando las 10 más correlacionadas a menos de 200 km, se desechan los datos con un percentil de diferencia de más de 0.6.}
\usage{
mediterranean_calculations(data, max_dist_eval)
}
\arguments{
\item{data}{ruta del fichero de datos}

\item{max_dist_eval}{máxima distancia entre 2 estaciones para usar una para evaluar una con la otra}
}
\value{
data y coor con los datos que no pasan el control eliminados
}
\description{
Hace el control de calidad
Control de calidad: Se estaciones con menos de 20 años de datos y usando las 10 más correlacionadas a menos de 200 km, se desechan los datos con un percentil de diferencia de más de 0.6.
}
