% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterraneancalculations.R
\docType{package}
\name{mediterraneancalculations-package}
\alias{mediterraneancalculations}
\alias{mediterraneancalculations-package}
\title{mediterraneancalculations: Mediterranean Calculations}
\description{
.
}
\details{
Info
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lcsc.csic.es}
}

}
