% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{mobile_trends}
\alias{mobile_trends}
\title{esto te calcula unas tendencias móviles de una serie, en este caso que empieza en 1851 y termina en 2018,. Habría que hacerlo para cada base de datos y estación.}
\usage{
mobile_trends(datos)
}
\arguments{
\item{datos}{datos}
}
\value{
list
}
\description{
esto te calcula unas tendencias móviles de una serie, en este caso que empieza en 1851 y termina en 2018,. Habría que hacerlo para cada base de datos y estación.
}
