% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{near_estations}
\alias{near_estations}
\title{Devuelve las estaciones por orden de cercanía
Sin tener en cuenta las que están a más de 200 km (NA en esos casos)}
\usage{
near_estations(data, coor, max_dist)
}
\arguments{
\item{data}{datos mensuales}

\item{coor}{coordenadas de las estaciones que corresponden con data}

\item{max_dist}{distancia máxima entre las series a utilizar}
}
\value{
correlación entre las estaciones
}
\description{
Devuelve las estaciones por orden de cercanía
Sin tener en cuenta las que están a más de 200 km (NA en esos casos)
}
