% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{order_data}
\alias{order_data}
\title{Sorts the data and returns a list with the order}
\usage{
order_data(data)
}
\arguments{
\item{data}{data}
}
\value{
list
}
\description{
Sorts the data and returns a list with the order
}
