% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{overlap_station}
\alias{overlap_station}
\title{Calcula el tiempo de solape existente entre cada par de series}
\usage{
overlap_station(control_data)
}
\arguments{
\item{control_data}{datos de las estaciones y sus coordenadas}
}
\value{
matriz con los meses que se solapan las estaciones entre si
}
\description{
Calcula el tiempo de solape existente entre cada par de series
}
