% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{percentage_of_zeros}
\alias{percentage_of_zeros}
\title{Devuelve el procentage de datos válidos que son 0s}
\usage{
percentage_of_zeros(data)
}
\arguments{
\item{data}{datos}
}
\value{
percentage
}
\description{
Devuelve el procentage de datos válidos que son 0s
}
