% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{read_data}
\alias{read_data}
\title{Leemos los datos desde los CSVs con el formato acordado}
\usage{
read_data(file_data, file_coor)
}
\arguments{
\item{file_data}{ruta del fichero de datos}

\item{file_coor}{ruta del fichero de coordenadas}
}
\value{
None
}
\description{
Leemos los datos desde los CSVs con el formato acordado
}
