% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{read_years}
\alias{read_years}
\title{Reads the years from text strings that end with years}
\usage{
read_years(txt)
}
\arguments{
\item{txt}{text or vector of texts}
}
\value{
list
}
\description{
Reads the years from text strings that end with years
}
