% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_csvs}
\alias{save_csvs}
\title{Guardamos los datos en CSVs}
\usage{
save_csvs(i_ini, folder_name, data_save, coor_save)
}
\arguments{
\item{i_ini}{identificador de los ficheros}

\item{folder_name}{carpeta en la que guardar el fichero}

\item{data_save}{datos de las estaciones a guardar}

\item{coor_save}{datos de coordenadas a guardar}
}
\value{
None
}
\description{
Guardamos los datos en CSVs
}
