% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_data}
\alias{save_data}
\title{Saves the output in 5 files with the data
 5 files indicating whether each data point is original or filled (1 for unaltered data, 0 for altered data)
 and 5 coordinate files for the stations in each data file, which are:
- 1870 to 2020 with more than 0.75 of years of original data
- 1900 to 2020 with more than 0.75 of years of original data
- 1930 to 2020 with more than 0.75 of years of original data
- 1950 to 2020 with more than 0.75 of years of original data
- 1990 to 2020 with more than 0.75 of years of original data}
\usage{
save_data(data_ori, control_data)
}
\arguments{
\item{data_ori}{original data read from CSV files}

\item{control_data}{data from the stations and their coordinates}
}
\value{
data and coordinates with data that did not pass the control removed
}
\description{
Saves the output in 5 files with the data
 5 files indicating whether each data point is original or filled (1 for unaltered data, 0 for altered data)
 and 5 coordinate files for the stations in each data file, which are:
- 1870 to 2020 with more than 0.75 of years of original data
- 1900 to 2020 with more than 0.75 of years of original data
- 1930 to 2020 with more than 0.75 of years of original data
- 1950 to 2020 with more than 0.75 of years of original data
- 1990 to 2020 with more than 0.75 of years of original data
}
