% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_delete_data}
\alias{save_delete_data}
\title{Para cada estación guardar, número de datos de entrada y de datos eliminados}
\usage{
save_delete_data(ori_data, process_data, folder)
}
\arguments{
\item{ori_data}{datos iniciales}

\item{process_data}{datos procesados}

\item{folder}{carpeta en la que se guarda el fichero resultante}
}
\value{
datos eliminados y datos de entrada por estación
}
\description{
Para cada estación guardar, número de datos de entrada y de datos eliminados
}
