% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_parameters}
\alias{save_parameters}
\title{Guardamos las variables globales utilizadas}
\usage{
save_parameters(folder_name, max_dist)
}
\arguments{
\item{folder_name}{carpeta en la que guardar el fichero}

\item{max_dist}{máxima distancia entre 2 estaciones para usar una para evaluar o completar la otra}
}
\value{
None
}
\description{
Guardamos las variables globales utilizadas
}
