% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{second_data_fill}
\alias{second_data_fill}
\title{Performs a second fill
For each dataset (1870, 1900...)
Stations with more than 90 or 95% of filled data
Stations are sorted by correlation (minimum 0.5), and filling is done using the 10 methods...
Stations without total fill are discarded}
\usage{
second_data_fill(data, max_dist_eval = NA)
}
\arguments{
\item{data}{data and coordinates}

\item{max_dist_eval}{maximum distance for filling}
}
\value{
data and coordinates
}
\description{
Performs a second fill
For each dataset (1870, 1900...)
Stations with more than 90 or 95% of filled data
Stations are sorted by correlation (minimum 0.5), and filling is done using the 10 methods...
Stations without total fill are discarded
}
