% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{second_data_fill_data}
\alias{second_data_fill_data}
\title{Realiza un segundo relleno
Para cada base de datos (1870, 1900...)
Estaciones con más del 90 o 95% de datos rellenos
Ordenamos las estaciones por correlación (mínima 0.5) y rellenamos usando los 10 métodos...
Las estaciones sin relleno total las tiramos}
\usage{
second_data_fill_data(file_data)
}
\arguments{
\item{file_data}{ruta del fichero de datos}
}
\value{
None
}
\description{
Realiza un segundo relleno
Para cada base de datos (1870, 1900...)
Estaciones con más del 90 o 95% de datos rellenos
Ordenamos las estaciones por correlación (mínima 0.5) y rellenamos usando los 10 métodos...
Las estaciones sin relleno total las tiramos
}
