% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{select_data}
\alias{select_data}
\title{Primeros datos válidos (no NAs)}
\usage{
select_data(data, n_reference_stations)
}
\arguments{
\item{data}{datos}

\item{n_reference_stations}{numero de datos a devolver}
}
\value{
primeros datos distintos de NA
}
\description{
Primeros datos válidos (no NAs)
}
