% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{sum_no_nas}
\alias{sum_no_nas}
\title{Number of non-NA data points}
\usage{
sum_no_nas(data)
}
\arguments{
\item{data}{data}
}
\value{
number of non-NA data points
}
\description{
Number of non-NA data points
}
