% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{alexanderson_homogenize}
\alias{alexanderson_homogenize}
\title{Test de Alexanderson para todos los ficheros disponibles (que han pasado el segundo relleno con éxito)}
\usage{
alexanderson_homogenize(data)
}
\arguments{
\item{data}{data y coor}
}
\value{
data y coor
}
\description{
Test de Alexanderson para todos los ficheros disponibles (que han pasado el segundo relleno con éxito)
}
