% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{alexanderson_homogenize_data}
\alias{alexanderson_homogenize_data}
\title{Homogeneizar - test de Alexanderson
Lo usamos en code_web_maps/snht_functions.R
Existe también librería snht de R
-- Serie de referencia, compara y corrige
Para cada base de datos (1870, 1900...)
Elegimos las 5 series más correlacionadas usandos las serie de diferencias
Con las 5 hacemos una media ponderada, (correlación * dato1 + ...) / sum(correlaciones) y será la serie de referencia
Alexanderson nos dará un punto de ruptura y un valor ratio por el que multiplicar la parte antigua... iterar mientras de puntos de ruptura
Guardar estadísticos de inhomogeneidades. Básicamente número de datos cambiados en cada series y momento de la inhomogeneidad. 
CSV con número de datos cambiados y CSV con punto de inhomogeneidad - todo x 12 meses}
\usage{
alexanderson_homogenize_data(file_data, no_use_series = c())
}
\arguments{
\item{file_data}{ruta del fichero de datos}

\item{no_use_series}{series que no se homogeneizarán}
}
\value{
None
}
\description{
Homogeneizar - test de Alexanderson
Lo usamos en code_web_maps/snht_functions.R
Existe también librería snht de R
-- Serie de referencia, compara y corrige
Para cada base de datos (1870, 1900...)
Elegimos las 5 series más correlacionadas usandos las serie de diferencias
Con las 5 hacemos una media ponderada, (correlación * dato1 + ...) / sum(correlaciones) y será la serie de referencia
Alexanderson nos dará un punto de ruptura y un valor ratio por el que multiplicar la parte antigua... iterar mientras de puntos de ruptura
Guardar estadísticos de inhomogeneidades. Básicamente número de datos cambiados en cada series y momento de la inhomogeneidad. 
CSV con número de datos cambiados y CSV con punto de inhomogeneidad - todo x 12 meses
}
