% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{apply_ecdf}
\alias{apply_ecdf}
\title{Monthly data anomalies}
\usage{
apply_ecdf(data)
}
\arguments{
\item{data}{monthly data}
}
\value{
data anomalies
}
\description{
Monthly data anomalies
}
