% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{apply_ecdf_month}
\alias{apply_ecdf_month}
\title{Data anomalies}
\usage{
apply_ecdf_month(data)
}
\arguments{
\item{data}{monthly data}
}
\value{
data anomalies
}
\description{
Data anomalies
}
