% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_data_year}
\alias{calc_data_year}
\title{Sums the data for each year to return a single annual value}
\usage{
calc_data_year(data)
}
\arguments{
\item{data}{data matrix}
}
\value{
one value per year
}
\description{
Sums the data for each year to return a single annual value
}
