% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_data_year_month_station}
\alias{calc_data_year_month_station}
\title{Devuelve el slope z por años y estaciones}
\usage{
calc_data_year_month_station(data, calc_function)
}
\arguments{
\item{data}{datos de las estaciones}

\item{calc_function}{función a utilizar}
}
\value{
lista de resultados
}
\description{
Devuelve el slope z por años y estaciones
}
