% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_mkTrend_pval}
\alias{calc_mkTrend_pval}
\title{Devuelve el pval calculado por mkTrend o el pval0 si el pval era NA}
\usage{
calc_mkTrend_pval(data)
}
\arguments{
\item{data}{matriz de datos}
}
\value{
pval
}
\description{
Devuelve el pval calculado por mkTrend o el pval0 si el pval era NA
}
