% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_mkTrend_slp}
\alias{calc_mkTrend_slp}
\title{regresi칩n lineal de los datos contra los a침os}
\usage{
calc_mkTrend_slp(data)
}
\arguments{
\item{data}{index}
}
\value{
lm
}
\description{
regresi칩n lineal de los datos contra los a침os
}
