% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calc_percentage}
\alias{calc_percentage}
\title{Diferencia en porcentaje}
\usage{
calc_percentage(datos, years = NA)
}
\arguments{
\item{datos}{datos}

\item{years}{años}
}
\value{
percentage
}
\description{
Diferencia en porcentaje
}
