% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{calculate_reconstruction_statistics}
\alias{calculate_reconstruction_statistics}
\title{Calcular estadísticos de la reconstrucción
- hydroGOF -- estadísitico por estación
- D / MAE / PBIAS / RMSE  - por estación y mes}
\usage{
calculate_reconstruction_statistics(sim, obs)
}
\arguments{
\item{sim}{datos rellenados}

\item{obs}{datos iniciales}
}
\value{
datos eliminados y datos de entrada por estación
}
\description{
Calcular estadísticos de la reconstrucción
- hydroGOF -- estadísitico por estación
- D / MAE / PBIAS / RMSE  - por estación y mes
}
