% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Salida final: serie regional promedio, tendencias, SPI...
Un RData o RDS y ficheros CSV sueltos por base da datos
Identificador, X, Y, series...}
\usage{
calculate_statistics(data)
}
\arguments{
\item{data}{data y coor}
}
\value{
data and coor
}
\description{
Salida final: serie regional promedio, tendencias, SPI...
Un RData o RDS y ficheros CSV sueltos por base da datos
Identificador, X, Y, series...
}
