% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Final output with all statistics, regional average series, trends, SPI...}
\usage{
calculate_statistics(data, data_ori)
}
\arguments{
\item{data}{data and coordinates}

\item{data_ori}{original data}
}
\value{
data and coordinates
}
\description{
Final output with all statistics, regional average series, trends, SPI...
}
