% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{calculate_statistics_data}
\alias{calculate_statistics_data}
\title{Calcula estadísticos de los datos
Tendencia mensual, estacional y anual, paquete Trend, función sens.slope
Sumar 1 a todo para evitar 0s
Significación, paquete modifiedmk, funsión bbsmk
Serie promedio de todo el país
SPI a escalas 3, 12, y 24 de cada serie, importante que sea imposible invertir las operaciones
Código Sergio para generar arrays y hacer figuras de tendencia}
\usage{
calculate_statistics_data(file_data, data_ori)
}
\arguments{
\item{file_data}{datos y coordenadas}

\item{data_ori}{datos originales}
}
\value{
None
}
\description{
Calcula estadísticos de los datos
Tendencia mensual, estacional y anual, paquete Trend, función sens.slope
Sumar 1 a todo para evitar 0s
Significación, paquete modifiedmk, funsión bbsmk
Serie promedio de todo el país
SPI a escalas 3, 12, y 24 de cada serie, importante que sea imposible invertir las operaciones
Código Sergio para generar arrays y hacer figuras de tendencia
}
