% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{coef_var}
\alias{coef_var}
\title{Coefficients of variation, standard deviation
https://fhernanb.github.io/Manual-de-R/varia.html}
\usage{
coef_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{data}

\item{na.rm}{Ignore NAs}
}
\value{
percentage
}
\description{
Coefficients of variation, standard deviation
https://fhernanb.github.io/Manual-de-R/varia.html
}
