% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{delete_zones}
\alias{delete_zones}
\title{Por debajo de 28 grados norte, eliminar estaciones}
\usage{
delete_zones(data)
}
\arguments{
\item{data}{data y coor}
}
\value{
data and coor
}
\description{
Por debajo de 28 grados norte, eliminar estaciones
}
