% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{dry_spell_trend}
\alias{dry_spell_trend}
\title{Calculates the trend. A 'years' object must be defined with the corresponding year for each case.}
\usage{
dry_spell_trend(index, threshold)
}
\arguments{
\item{index}{index}

\item{threshold}{threshold}
}
\value{
output
}
\description{
Calculates the trend. A 'years' object must be defined with the corresponding year for each case.
}
