% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{fill_one_series}
\alias{fill_one_series}
\title{Fills the received series using others in the order they appear in other_series}
\usage{
fill_one_series(series, other_series)
}
\arguments{
\item{series}{data series to complete}

\item{other_series}{data series to use for completing in the order they must be used}
}
\value{
filled data series
}
\description{
Fills the received series using others in the order they appear in other_series
}
