% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{fill_series}
\alias{fill_series}
\title{Rellenado mensual de las series
Usamos estaciones a menos de 200km con correlación por encima de 0.7
Para junio, julio y agosto, rellenamos con la más cercana
Utilizar el método que mejor correlaciona con la serie original}
\usage{
fill_series(control_data, min_correlation, max_dist)
}
\arguments{
\item{control_data}{datos de las estaciones y sus coordenadas}

\item{min_correlation}{Correlación mínima para usar el dato en el relleno}

\item{max_dist}{distancia máxima entre las series a utilizar}
}
\value{
data y coor con los datos que no pasan el control eliminados
}
\description{
Rellenado mensual de las series
Usamos estaciones a menos de 200km con correlación por encima de 0.7
Para junio, julio y agosto, rellenamos con la más cercana
Utilizar el método que mejor correlaciona con la serie original
}
