% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{main_mediterranean_calculations}
\alias{main_mediterranean_calculations}
\title{Lee los ficheros de precipitación, calcula estadísticos y guarda los resultados
Los ficheros de entrada son 2 CSVs uno de coordenadas en grados (filas las estaciones y columnas lat y lon y otro de datos mensuales con fechas en filas y las estaciones en las columnas)}
\usage{
main_mediterranean_calculations(file_data, file_coor)
}
\arguments{
\item{file_data}{ruta del fichero de datos}

\item{file_coor}{ruta del fichero de coordenadas}
}
\value{
None
}
\description{
Lee los ficheros de precipitación, calcula estadísticos y guarda los resultados
Los ficheros de entrada son 2 CSVs uno de coordenadas en grados (filas las estaciones y columnas lat y lon y otro de datos mensuales con fechas en filas y las estaciones en las columnas)
}
