% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{main_mediterranean_calculations_}
\alias{main_mediterranean_calculations_}
\title{Calcula los estadísticos para un país}
\usage{
main_mediterranean_calculations_(read_all_data, folder, pb = NULL)
}
\arguments{
\item{read_all_data}{datos de entrada}

\item{folder}{carpeta donde guarda ficheros}

\item{pb}{barra de progreso}
}
\value{
None
}
\description{
Calcula los estadísticos para un país
}
