% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{mediterranean_calculations}
\alias{mediterranean_calculations}
\title{Performs quality control
Quality control: Stations with less than 20 years of data are removed, and using the 10 most correlated stations within 200 km, data with a percentile difference greater than 0.6 are discarded.}
\usage{
mediterranean_calculations(data, max_dist_eval)
}
\arguments{
\item{data}{path to the data file}

\item{max_dist_eval}{maximum distance between two stations to use one to evaluate or complete the other}
}
\value{
data and coordinates with data that do not pass the control removed
}
\description{
Performs quality control
Quality control: Stations with less than 20 years of data are removed, and using the 10 most correlated stations within 200 km, data with a percentile difference greater than 0.6 are discarded.
}
