% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{mediterranean_calculations}
\alias{mediterranean_calculations}
\title{Lee los ficheros de precipitación, hace el relleno y guarda los ficheros de salida
Los ficheros de entrada son 2 CSVs uno de coordenadas en grados (filas las estaciones y columnas lat y lon y otro de datos mensuales con fechas en filas y las estaciones en las columnas)
Control de calidad: Se estaciones con menos de 20 años de datos y usando las 10 más correlacionadas a menos de 200 km, se desechan los datos con un percentil de diferencia de más de 0.6.
Rellenado mensual de las series: Usamos estaciones a menos de 200km con correlación por encima de 0.7. Excepto para junio, julio y agosto, meses para los que rellenamos con la más cercana.
Homogeneiza, test de Alexandersson (snht)
Guarda la salida en 5 fichero con los datos
 5 ficheros que indican si cada datos es original o rellenado 
 y 5 ficheros de coordenadas para las estaciones de cada fichero de datos, que son:
- 1870 a 2020 con más de 80 años originales
- 1900 a 2020 con más de 80 años originales
- 1930 a 2020 con más de 60 años originales
- 1950 a 2020 con más de 40 años originales
- 1990 a 2020 con más de 30 años originales}
\usage{
mediterranean_calculations(
  file_data = "data.csv",
  file_coor = "coor.csv",
  max_dist
)
}
\arguments{
\item{file_data}{ruta del fichero de datos}

\item{file_coor}{ruta del fichero de coordenadas}

\item{max_dist}{máxima distancia entre 2 estaciones para usar una para evaluar o completar la otra}
}
\value{
data y coor con los datos que no pasan el control eliminados
}
\description{
Lee los ficheros de precipitación, hace el relleno y guarda los ficheros de salida
Los ficheros de entrada son 2 CSVs uno de coordenadas en grados (filas las estaciones y columnas lat y lon y otro de datos mensuales con fechas en filas y las estaciones en las columnas)
Control de calidad: Se estaciones con menos de 20 años de datos y usando las 10 más correlacionadas a menos de 200 km, se desechan los datos con un percentil de diferencia de más de 0.6.
Rellenado mensual de las series: Usamos estaciones a menos de 200km con correlación por encima de 0.7. Excepto para junio, julio y agosto, meses para los que rellenamos con la más cercana.
Homogeneiza, test de Alexandersson (snht)
Guarda la salida en 5 fichero con los datos
 5 ficheros que indican si cada datos es original o rellenado 
 y 5 ficheros de coordenadas para las estaciones de cada fichero de datos, que son:
- 1870 a 2020 con más de 80 años originales
- 1900 a 2020 con más de 80 años originales
- 1930 a 2020 con más de 60 años originales
- 1950 a 2020 con más de 40 años originales
- 1990 a 2020 con más de 30 años originales
}
