% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterraneancalculations.R
\docType{package}
\name{mediterraneancalculations-package}
\alias{mediterraneancalculations}
\alias{mediterraneancalculations-package}
\title{mediterraneancalculations: Mediterranean Calculations}
\description{
A suite of functions designed for quality control, reconstruction, and homogenization of monthly precipitation series, enabling the generation of series for specific periods and the analysis of precipitation trends.
}
\details{
Function \code{\link{main_mediterranean_calculations}} are the workhorse of the mediterraneancalculations library. Other functions such as \code{\link{calculate_statistics_data}}, \code{\link{alexanderson_homogenize}} is an auxiliary low-level function and will not be used directly by the typical user.
}
\references{
\url{https://journals.ametsoc.org/view/journals/clim/32/22/jcli-d-19-0244.1.xml} \url{https://rmets.onlinelibrary.wiley.com/doi/10.1002/joc.2115}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lcsc/mediterraneancalculations}
  \item \url{https://lcsc.csic.es}
}

}
\author{
Sergio M. Vicente-Serrano and Fergus Reig-Gracia
}
