% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{mkTrend}
\alias{mkTrend}
\title{Calculates p-value (sometimes it doesn't result due to iteration issues, so take pval0).}
\usage{
mkTrend(x, ci = 0.95)
}
\arguments{
\item{x}{x}

\item{ci}{ci}
}
\value{
list
}
\description{
Calculates p-value (sometimes it doesn't result due to iteration issues, so take pval0).
}
