% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{mkTrend}
\alias{mkTrend}
\title{esta función calcula las tendencias. Lo que nos interesa es slp (es la función de cálculo del Sen slope), pval (a veces no da resultado por temas de iteración) entonces coger pval0.}
\usage{
mkTrend(x, ci = 0.95)
}
\arguments{
\item{x}{x}

\item{ci}{ci}
}
\value{
list
}
\description{
esta función calcula las tendencias. Lo que nos interesa es slp (es la función de cálculo del Sen slope), pval (a veces no da resultado por temas de iteración) entonces coger pval0.
}
