% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{mobile_trends}
\alias{mobile_trends}
\title{Calculates the moving trends of a series.}
\usage{
mobile_trends(datos)
}
\arguments{
\item{datos}{datos}
}
\value{
list
}
\description{
Calculates the moving trends of a series.
}
