% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{near_correlations}
\alias{near_correlations}
\title{Returns the correlation between stations
Ignoring those that are more than 200 km away (NA in those cases)}
\usage{
near_correlations(data, coor, max_dist)
}
\arguments{
\item{data}{monthly data}

\item{coor}{coordinates of the stations corresponding to the data}

\item{max_dist}{maximum distance between the series to be used}
}
\value{
correlation between stations
}
\description{
Returns the correlation between stations
Ignoring those that are more than 200 km away (NA in those cases)
}
