% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{order_data}
\alias{order_data}
\title{Ordena los datos y devuelve una lista con el orden}
\usage{
order_data(data)
}
\arguments{
\item{data}{datos}
}
\value{
list
}
\description{
Ordena los datos y devuelve una lista con el orden
}
