% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{overlap_station}
\alias{overlap_station}
\title{Calculates the overlap time between each pair of series}
\usage{
overlap_station(control_data)
}
\arguments{
\item{control_data}{data from the stations and their coordinates}
}
\value{
matrix with the months where the stations overlap with each other
}
\description{
Calculates the overlap time between each pair of series
}
