% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{overlap_station_no_0}
\alias{overlap_station_no_0}
\title{Calculates the overlap time between each pair of series without counting zeros}
\usage{
overlap_station_no_0(control_data)
}
\arguments{
\item{control_data}{data from the stations and their coordinates}
}
\value{
matrix with the months where the stations overlap with each other
}
\description{
Calculates the overlap time between each pair of series without counting zeros
}
