% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{percentage_of_zeros}
\alias{percentage_of_zeros}
\title{Returns the percentage of valid data that are zeros}
\usage{
percentage_of_zeros(data)
}
\arguments{
\item{data}{data}
}
\value{
percentage
}
\description{
Returns the percentage of valid data that are zeros
}
