% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{quality_control}
\alias{quality_control}
\title{Quality control
Stations with less than 20 years of data are removed
Using the 10 closest within 200 km, discard if the average percentile differs by more than 0.6
or more than 0.5 for data points with a value of 0}
\usage{
quality_control(data, coor, max_dist, max_diff_anomaly, max_diff_anomaly_0)
}
\arguments{
\item{data}{data}

\item{coor}{coordinates}

\item{max_dist}{maximum distance between two stations to use one to evaluate or complete the other}

\item{max_diff_anomaly}{maximum anomaly difference to keep data in quality control}

\item{max_diff_anomaly_0}{maximum anomaly difference to keep data in quality control, if the data point is 0}
}
\value{
data and coor with data points that do not pass the control removed
}
\description{
Quality control
Stations with less than 20 years of data are removed
Using the 10 closest within 200 km, discard if the average percentile differs by more than 0.6
or more than 0.5 for data points with a value of 0
}
