% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{read_data}
\alias{read_data}
\title{Reads data from CSVs in the agreed format
The input files are 2 CSVs: one with coordinates in degrees (stations in rows and columns lat and lon) and another with monthly data (dates in rows and stations in columns)}
\usage{
read_data(file_data, file_coor)
}
\arguments{
\item{file_data}{path to the data file}

\item{file_coor}{path to the coordinates file}
}
\value{
original data, data of interest, and coordinates of the stations read
}
\description{
Reads data from CSVs in the agreed format
The input files are 2 CSVs: one with coordinates in degrees (stations in rows and columns lat and lon) and another with monthly data (dates in rows and stations in columns)
}
