% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{read_years}
\alias{read_years}
\title{Lee los a単os de cadenas de texto que terminan con los a単os}
\usage{
read_years(txt)
}
\arguments{
\item{txt}{texto o vector de textos}
}
\value{
list
}
\description{
Lee los a単os de cadenas de texto que terminan con los a単os
}
