% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_csvs}
\alias{save_csvs}
\title{Saves the data into CSVs}
\usage{
save_csvs(i_ini, folder_name, data_save, coor_save)
}
\arguments{
\item{i_ini}{identifier for the files}

\item{folder_name}{folder where the file will be saved}

\item{data_save}{data from the stations to save}

\item{coor_save}{coordinates data to save}
}
\value{
None
}
\description{
Saves the data into CSVs
}
