% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_data}
\alias{save_data}
\title{Guarda la salida en 5 fichero con los datos
 5 ficheros que indican si cada datos es original o rellenado (1 dato no alterado, 0 dato alterado)
 y 5 ficheros de coordenadas para las estaciones de cada fichero de datos, que son:
- 1870 a 2020 con más de 80 años originales
- 1900 a 2020 con más de 80 años originales
- 1930 a 2020 con más de 60 años originales
- 1950 a 2020 con más de 40 años originales
- 1990 a 2020 con más de 30 años originales}
\usage{
save_data(data_ori, control_data)
}
\arguments{
\item{data_ori}{datos originales leidos de los ficheros CSV}

\item{control_data}{datos de las estaciones y sus coordenadas}
}
\value{
data y coor con los datos que no pasan el control eliminados
}
\description{
Guarda la salida en 5 fichero con los datos
 5 ficheros que indican si cada datos es original o rellenado (1 dato no alterado, 0 dato alterado)
 y 5 ficheros de coordenadas para las estaciones de cada fichero de datos, que son:
- 1870 a 2020 con más de 80 años originales
- 1900 a 2020 con más de 80 años originales
- 1930 a 2020 con más de 60 años originales
- 1950 a 2020 con más de 40 años originales
- 1990 a 2020 con más de 30 años originales
}
