% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{save_delete_data}
\alias{save_delete_data}
\title{For each station, save the number of input data and deleted data}
\usage{
save_delete_data(ori_data, process_data, folder)
}
\arguments{
\item{ori_data}{original data}

\item{process_data}{processed data}

\item{folder}{folder where the resulting file is saved}
}
\value{
deleted data and input data per station
}
\description{
For each station, save the number of input data and deleted data
}
