% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediterranean-calculations.r
\name{second_data_fill}
\alias{second_data_fill}
\title{Realiza un segundo relleno
Para cada base de datos (1870, 1900...)
Estaciones con más del 90 o 95% de datos rellenos
Ordenamos las estaciones por correlación (mínima 0.5) y rellenamos usando los 10 métodos...
Las estaciones sin relleno total las tiramos}
\usage{
second_data_fill(data)
}
\arguments{
\item{data}{data y coor}
}
\value{
data y coor
}
\description{
Realiza un segundo relleno
Para cada base de datos (1870, 1900...)
Estaciones con más del 90 o 95% de datos rellenos
Ordenamos las estaciones por correlación (mínima 0.5) y rellenamos usando los 10 métodos...
Las estaciones sin relleno total las tiramos
}
