% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{select_data}
\alias{select_data}
\title{First valid data points (non-NAs)}
\usage{
select_data(data, n_reference_stations)
}
\arguments{
\item{data}{data}

\item{n_reference_stations}{number of data points to return}
}
\value{
first non-NA data points
}
\description{
First valid data points (non-NAs)
}
