% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mediterranean-calculations.r
\name{sum_no_nas}
\alias{sum_no_nas}
\title{Número de datos distintos de NA}
\usage{
sum_no_nas(data)
}
\arguments{
\item{data}{datos}
}
\value{
número de datos no NAs
}
\description{
Número de datos distintos de NA
}
